==> Instacart_Basket_Analysis_2023_04_03/03_Scripts/01_improvised_data_dictionary_and_data_transformation.ipynb <==
## Contents
1. Importing libraries
2. Reading data, setting up
3. Dropping and renaming columns
4. Changing datatypes, reading and transposing department data
5. Creating data dictionary and snacks subset
6. Transforming data, creating data dictionary

### 1. Importing Libraries

# import libraries used
import pandas as pd
import numpy as np
import os

### 2. Read orders and products data

# establish file pathto data, then transforming csv into dataframe
file_path = r'C:\Users\shurg\OneDrive\Asztali gép\CareerFoundry Data Analytics\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data\Original_Data'
df_orders = pd.read_csv(os.path.join(file_path, 'orders.csv'), index_col = False)
df_products = pd.read_csv(os.path.join(file_path, 'products.csv'), index_col = False)


### 3. Dropping and renaming columns

# removing column eval_set
df_orders = df_orders.drop(columns = ['eval_set'])

df_orders['days_since_prior_order'].value_counts(dropna = False)

# rename dow
df_orders.rename(columns = {'order_dow' : 'orders_day_of_the_week'}, inplace = True)
df_orders.head()

### 4. Changing datatypes, reading and transposing department data

#change order_id datatype to string
df_orders['order_id'] = df_orders['order_id'].astype('str')
df_orders['order_id'].dtype

# read and transpose departments data
df_departments = pd.read_csv(os.path.join(file_path, 'departments.csv'), index_col = False)
df_departments_t = df_departments.T
df_departments_t = df_departments_t.reset_index()
df_departments_t.head()
new_header = df_departments_t.iloc[0]
df_depts = df_departments_t[1:]
df_depts.columns = new_header
df_depts


### 5. Creating data dictionary and snacks subset

# creating a data dictionary from departments data
data_dict = df_depts.to_dict('index')
data_dict

# creating a subset of products 'snacks'
df_snacks = df_products[df_products['department_id'] == 19]
df_snacks

### 6. Transforming data, creating data dictionary

# Task 2.
df_orders['user_id'] = df_orders['user_id'].astype('str')
df_orders['user_id'].dtype

# Task 3.
df_orders.rename(columns = {'days_since_prior_order' : 'days_after_previous_order'}, inplace = True)
df_orders.head()

# Task 4. 
df_orders['order_hour_of_day'].value_counts(dropna = False)
# the most busy hour is 10, apperaring 288418 times

# Task 5.
print(data_dict)
# Answer: the departmen name is 'produce'

# Task 6.
df_breakfast = df_products[df_products['department_id'] == 14]
df_breakfast.head()

# Task 7 
df_dinnerparty = df_products[df_products['department_id'].isin([5, 7, 12, 20])]
df_dinnerparty.head()

# Task 8
len(df_dinnerparty)
# Answer: There are 7650 rows

# Task9
user_no1 = df_orders[df_orders['user_id'] == '1']
print(user_no1)

# Task 10 
# Answers: 
# - Made 11 orders on instacart
# - Makes orders on Munday, Tuesday, Wednesday, and Thurstday
# - Mostly makes orders in the morning (7-9) or in the afternoon (12-16)
# - Order frequency varies quite a bit, betwwen 0 an 30 days

# Task 11, export to csv
save_path = r'C:\Users\shurg\OneDrive\Asztali gép\CareerFoundry Data Analytics\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data\Prepared_Data'
df_orders.to_csv(os.path.join(save_path, 'orders_wrangled.csv'))
df_depts.to_csv(os.path.join(save_path, 'departments.wrangled.csv'))
