==> Instacart_Basket_Analysis_2023_04_03/03_Scripts/02_cleaning_and_consistency_checks.ipynb <==
## Contents
1. Importing libraries
2. Reading data, setting up
3. Testing mixed datatypes
4. Products data consistency checks
5. Orders data consistency checks

### Importing Libraries

# importing libraries
import pandas as pd
import numpy as np
import os

### 2. Reading data, setting up

# reading products and orders data
path_org = r'C:\Users\Richárd\Desktop\CareerFoundry\Data Immersion\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data\Original_Data'
path_prep = r'C:\Users\Richárd\Desktop\CareerFoundry\Data Immersion\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data\Prepared_Data'
df_prods = pd.read_csv(os.path.join(path_org, 'products.csv'), index_col = None)
df_ords = pd.read_csv(os.path.join(path_prep, 'orders_wrangled.csv'), index_col = None)

### 3. Testing mixed datatypes

# setting up test dataframe
df_test = pd.DataFrame()
df_test['mix'] = ['a', 'b', 1, True]
df_test.head()

# looping through test dataframe columns [converted to lists], mapping them for different datatypes
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# converting mix column to str
df_test['mix'] = df_test['mix'].astype('str')

### 4. Products data consistency checks

# summing up null values
df_prods.isnull().sum()

# looking at only null values
df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

df_prods.shape

# filtering out null values, assigning them to a new variable
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]
df_prods_clean.shape

# identifying duplicates, assigning them to a variable
df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

# Getting rid of duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups.shape

### 5. Orders data consistency checks

# Task 1. - Please see above

# Task 2. 
df_ords.describe() # - nothing out of the ordinary here

# Task2+ products data was never checked
df_prods_clean_no_dups.describe() # - not so cean after all! 99999.000000 for max price seems quite extreme, wort looking into

# Task 3. - Checking for mixed datatypes in orders data
for col in df_ords.columns.tolist():
    is_mixed = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    if len(df_ords[is_mixed]) > 0:
        print(col)
# Task 4. - nothing gets printed, meaning no mixed dataype columns - Nothing to fix
    

# Task 5. Checking for missing values in orders data
orders_nulls_summary = df_ords.isnull().sum()
orders_nulls_summary # - days_after_previous_order has 206209 null values - likely means that 0 days passed between orders, worth taking a closer look though

# exploring reason behind missing values
ords_nulls = df_ords[df_ords['days_after_previous_order'].isnull() == True]
# ords_nulls
exmpl = df_ords[df_ords['order_number'] == 1]
exmpl.isnull().sum()


#### Task 5. explaination: Clearly the missing value indicate the first order made by each user - no days could have passed if there is no previous order

#### Task 6. It could be reasonable to flag these as something like 'first_order_of_user', but then we would be dealing with mixed datatype for the column. Replacing them with 0-s would be a bad idea, because it would skew analysis of order frequencies.
#### I think the best solution is to leave them be, but mark it down in the docuentation/data dictionary that nulls mean the very first order made by the user.

# Task 7. - Finding duplicates in orders data
ords_dups = df_ords[df_ords.duplicated()]
ords_dups

df_prods.shape

df_ords.shape

#### Task 7., Task 8. - There appear to be no duplicates, so nothing to do here

# Task 9. - Exporting dataframes to csvs
df_prods_clean_no_dups.to_csv(os.path.join(path_prep, 'products_clean.csv'))
df_ords .to_csv(os.path.join(path_prep, 'orders_clean.csv'))
