==> Instacart_Basket_Analysis_2023_04_03/03_Scripts/03_merge_orders_and_product.ipynb <==
### Import Libraries

# import libraries
import pandas as pd
import numpy as np
import os

### Import data

path = r'C:\Users\Richárd\Desktop\CareerFoundry\Data Immersion\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data'
df_ords = pd.read_csv(os.path.join(path, 'Prepared_Data', 'orders_clean.csv'), index_col = False)
df_ords_prior = pd.read_csv(os.path.join(path, 'Original_Data', 'order_products__prior.csv'))

### Checking DataFrames

df_ords.head

df_ords_prior.head

df_ords.shape

df_ords_prior.shape

### Merge the two datasets

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large['_merge'].value_counts()

df_merged_large.to_pickle(os.path.join(path, 'Prepared_Data', 'orders_products_combined.pkl'))
