==> Instacart_Basket_Analysis_2023_04_03/03_Scripts/04_merging_orders_products_v2.ipynb <==
## Contents
1. Importing libraries
2. Reading data, setting up
3. Merging and exporting datasets

### 1. Importing libraries

# importing libraries
import pandas as pd
import numpy as np
import os

### 2. Reading Data, setting up

# Task 3.
# Importing data from pickle + product data from csv
path = R'C:\Users\Richárd\Desktop\CareerFoundry\Data Immersion\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data\Prepared_Data'
orders_products_combined = pd.read_pickle(os.path.join(path, 'orders_products_combined.pkl'))
orders_products_combined = orders_products_combined.filter(['order_id', 'user_id', 'order_number', 'orders_day_of_the_week', 'order_hour_of_day', 'days_after_previous_order', 'product_id', 'add_to_cart_order', 'reordered'])
df_products = pd.read_csv(os.path.join(path, 'products_clean.csv'),usecols = ['product_id', 'product_name', 'aisle_id', 'department_id', 'prices'], index_col = False)

# Task 4. - Checking shape of dataframe
orders_products_combined.shape

df_products.head()

orders_products_combined.head()

### 3. Merging and exporting datasets

# Task 5. Join the dataframes
df_merged = orders_products_combined.merge(df_products, on = 'product_id', how = 'inner', indicator = True)

df_merged.head()

# Task 6. - Confirm results using the merge flag
df_merged['_merge'].value_counts()

# Task 7. Export merged data to pickle
df_merged.to_pickle(os.path.join(path, 'ords_prods_merged_final.pkl'))
