==> Instacart_Basket_Analysis_2023_04_03/03_Scripts/05_create_initial_flag_columns.ipynb <==
## Contents
1. Importing libraries
2. Reading data, setting up
3. Creating flags and flag columns

### 1. Importing libraries

# importing libraries
import pandas as pd
import numpy as np
import os

### 2. Reading data

# Read data from pickle
path = r'C:\Users\Richárd\Desktop\CareerFoundry\Data Immersion\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data\Prepared_Data'
merged_data = pd.read_pickle(os.path.join(path, 'ords_prods_merged_final.pkl'))
# creating subset
df = merged_data[:1000000]
df.head()

### 3. Creating flags and flag columns

### Creating function to flag price-ranges

# Function to assign price_range labels
def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif row['prices'] > 5 and row['prices'] <= 15:
        return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High-range product'
    else:
        return 'Not enough data'

### Creating flag column

# Applying price_label() to df
df['price_range'] = df.apply(price_label, axis = 1)

# Frequency of price range labels
df['price_range'].value_counts()

# checking out the maximum
df['prices'].max()

### Using loc to assign price ranges

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df.loc[(df['prices'] > 5) & (df['prices'] <= 15), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Frequency of price ranges
df['price_range_loc'].value_counts()

### Assigning price-ranges to the whole merged dataset
#### Assigning flags using .loc

merged_data.loc[merged_data['prices'] > 15, 'price_range'] = 'High-range product'

merged_data.loc[(merged_data['prices'] > 5) & (merged_data['prices'] <= 15), 'price_range'] = 'Mid-range product'

merged_data.loc[merged_data['prices'] <= 5, 'price_range'] = 'Low-range product'

# Frequency of price ranges
merged_data['price_range'].value_counts()

merged_data['orders_day_of_the_week'].value_counts()

# looping to flag days based on busyness, adding them to a list
result = []

for value in merged_data['orders_day_of_the_week']:
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy day')
    else:
        result.append('Regularly busy')


# Adding 'busiest_day' column
merged_data['busiest_day'] = result
# Frequency of 'busiest_day' column
merged_data['busiest_day'].value_counts()

## Tasks

# Task 2. Flagging busiest days
days = []

for day in merged_data['orders_day_of_the_week']:
    if day == 0 or day == 1:
        days.append('Busiest days')
    elif day == 4 or day == 3:
        days.append('Least busy days')
    else:
        days.append('Regularly busy')

# Creating 'busiest_days' column
merged_data['busiest_days'] = days
merged_data['busiest_days'].value_counts()

### Task 3.
The two busiest days make up a disproportionately large number of orders it seems, there are significantly more orderder on those days.

# Task 4. identifying busiest hours
merged_data['order_hour_of_day'].value_counts()


# Setting up hour groups, flagging hours by number of orders
most_orders = [10, 11, 14, 15, 13, 12, 16, 9]
average_orders = [17, 8, 18, 19, 20, 7, 21, 22]

hours = []

for hour in merged_data['order_hour_of_day']:
    if hour in most_orders:
        hours.append('Most orders')
    elif hour in average_orders:
        hours.append('Average orders')
    else:
        hours.append('Fewest orders')

# Creating 'busiest_period_of_day' column
merged_data['busiest_period_of_day'] = hours
# Task 5. Displaying Frequency
merged_data['busiest_period_of_day'].value_counts()

# Task 6 exporting df to pickle
merged_data.to_pickle(os.path.join(path, 'merged_flagged_data.pkl'))
