==> Instacart_Basket_Analysis_2023_04_03/03_Scripts/07_clean_wrangle_merge_customer_data.ipynb <==
## Contents
1. Importing libraries
2. Reading data, setting up
3. Data wrangling and cleaning
4. Consistency check & cleansing
5. Combining datasets

### 1. Importing libraries

# importing libraries
import pandas as pd
import numpy as np
import os

### 2. Setup

# Task 3.
# setting up file paths, reading data
path_original = r'C:\Users\Richárd\Desktop\CareerFoundry\Data Immersion\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data\Original_Data'
path_prepared = r'C:\Users\Richárd\Desktop\CareerFoundry\Data Immersion\Archievment 4\Instacart_Basket_Analysis_2023_04_03\02_Data\Prepared_Data'
df_customers = pd.read_csv(os.path.join(path_original, 'customers.csv'))
df_customers.head()

### 3. Data wrangling and cleaning

# Task 4. Wrangling
# consistently rename columns
df_customers = df_customers.rename(columns = {'First Name': 'first_name', 'Surnam': 'surname', 'Gender': 'gender', 'STATE': 'state', 'Age': 'age', 'n_dependants': 'number_of_dependants', 'fam_status': 'family_status'})
# There seems to be no reason to include actual name s of the customers - moreover it raises a privacy issue
# creating a subset without first_name and surname columns
df_cust = df_customers.drop(columns = ['first_name', 'surname'])
# checking descriptive statistics
df_cust.head()

### 4. Consistency checks and cleansing

# Task 5. Consistency checks
# checking decriptive statistics
df_cust.describe() # everything seems reasonable here

# checking datatype consistency
for col in df_cust:
  weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)
# no mixed type columns found

# checking for duplicates
dups = df_cust.duplicated()
dups.value_counts() # there seem to be no duplicates

# checking for any null values
df_cust.isnull().sum() # no null values found

###  5. Combining datasets

# Task 6. Combining customer data with the main dataset
# reading data
main_data = pd.read_pickle(os.path.join(path_prepared, 'merged_data_with_groupings.pkl'))
main_data.head()

# joining the datasets
merged_data = main_data.merge(df_cust, on = 'user_id', how = 'inner')
merged_data.head()

# Exporting combined dataset to a pickle
merged_data.to_pickle(os.path.join(path_prepared, 'full_data.pkl'))
